requirements-autodl.txt
accelerate>=0.34.0
datasets>=2.20.0
huggingface_hub>=0.24.0
jsonlines>=4.0.0
numpy>=1.26.0
openai>=1.35.0
peft>=0.14.0
PyYAML>=6.0.1
sentence-transformers>=3.0.1
torch>=2.3.0
tqdm>=4.66.0
transformers>=4.46.0
trl>=0.15.2
